+++ gdb/breakpoint/int3.c
@@ -1,1 +1,1 @@
-insn.c
+breakpoint.c